[modify] modify the structure of original handlers
the master file is my_ruboty.rb and each handler is in original_handlers/
